split api tests to separate files for quicker build
