+++ backend/bin/__test__/syncTMCUsers.test.ts
@@ -4,8 +4,8 @@ import {
   fakeGetAccessToken,
   fakeUserDetailReply,
   getTestContext,
-} from "../../tests/__helpers"
-import { seed } from "../../tests/data/seed"
+} from "../../tests"
+import { seed } from "../../tests/data"
 import { TMCError } from "../lib/errors"
 import {
   Change,
