+++ backend/bin/kafkaConsumer/common/userCourseProgress/__test__/saveToDB.test.ts
@@ -1,5 +1,5 @@
-import { getTestContext } from "../../../../../tests/__helpers"
-import { seed } from "../../../../../tests/data/seed"
+import { getTestContext } from "../../../../../tests"
+import { seed } from "../../../../../tests/data"
 import { KafkaContext } from "../../kafkaContext"
 import { Message } from "../interfaces"
 import { saveToDatabase } from "../saveToDB"
