+++ backend/bin/kafkaConsumer/common/userPoints/__test__/saveToDB.test.ts
@@ -4,9 +4,12 @@ import {
   fakeGetAccessToken,
   fakeTMCSpecific,
   getTestContext,
-} from "../../../../../tests/__helpers"
-import { adminUserDetails, normalUserDetails } from "../../../../../tests/data"
-import { seed } from "../../../../../tests/data/seed"
+} from "../../../../../tests"
+import {
+  adminUserDetails,
+  normalUserDetails,
+  seed,
+} from "../../../../../tests/data"
 import { DatabaseInputError } from "../../../../lib/errors"
 import { KafkaContext } from "../../kafkaContext"
 import { Message } from "../interfaces"
