+++ frontend/components/Dashboard/Editor2/Common/Fields/ControlledModuleList.tsx
@@ -18,6 +18,7 @@ import {
   ListItem,
 } from "@mui/material"
 
+import { FormValues } from "../../types"
 import { EnumeratingAnchor } from "/components/Dashboard/Editor2/Common"
 import {
   ControlledFieldProps,
@@ -39,7 +40,9 @@ interface ControlledModuleListProps extends ControlledFieldProps {
   modules?: StudyModuleDetailedFieldsFragment[]
 }
 
-export function ControlledModuleList<T>(props: ControlledModuleListProps) {
+export function ControlledModuleList<T extends FormValues>(
+  props: ControlledModuleListProps,
+) {
   const { modules, label } = props
   const name = props.name as Path<T>
   const { setValue, getValues } = useFormContext<T>()
